This has been tested on Thursday at south florida. A couple changes were made which must be tested immediately Friday morning.

+++ Production/AerialPig2014/src/edu/wpi/first/wpilibj/templates/UserMessages.java
@@ -23,7 +23,7 @@ public class UserMessages {
         errorMessage = new StringBuffer("Everything is OK");
     }
     
-    public void setArmMessage(int armAngle, boolean charged, boolean ball)
+    public void setArmMessage(double armAngle, boolean charged, boolean ball)
     {
         armMessage = new StringBuffer("Arm: ").append(armAngle);
